fix: enforce recent memory retrieval scope
