+++ web/src/components/ai/AiMessageStatus.tsx
@@ -19,7 +19,16 @@ function buildScopeSummary(plan: PlannerAgentDto, t: ReturnType<typeof useTransl
   if (!scope) return summary;
   const listSeparator = t('scope.listSeparator');
 
-  if (scope.timeRange) summary.push(t('scope.time', { label: scope.timeRange.label }));
+  if (scope.selection === 'recent') {
+    summary.push(
+      t('scope.latestItems', {
+        count: scope.limit,
+        field: t(`scope.dateFields.${scope.dateField}`),
+      })
+    );
+  } else if (scope.timeRange) {
+    summary.push(t('scope.time', { label: scope.timeRange.label }));
+  }
 
   if (scope.tags.length) {
     summary.push(t('scope.tags', { tags: scope.tags.map((tag) => `#${tag}`).join(listSeparator) }));
