+++ web/src/locales/ja.json
@@ -657,6 +657,11 @@
         "title": "範囲",
         "paginationMore": "さらに表示",
         "time": "時間：{{label}}",
+        "latestItems": "最新{{count}}件（{{field}}）",
+        "dateFields": {
+          "createdAt": "作成日時",
+          "updatedAt": "更新日時"
+        },
         "tags": "タグ：{{tags}}",
         "excludeTags": "除外タグ：{{tags}}",
         "semanticScope": "キーワード：{{keywords}}",
