+++ web/src/utils/aiApi.ts
@@ -1,228 +1,27 @@
 import { authService } from './auth';
 import { get, getApiUrl, post, refreshAccessToken } from './api';
+import type {
+  AiAgentClientState,
+  AiAgentPhase,
+  AiAgentToolProgressStatus,
+  AiChatPayload,
+  AiChatStreamHandlers,
+  AiClientRequestContext,
+  AiClarification,
+  AiProviderTestProgressHandler,
+  AiProviderTestResult,
+  AiSemanticResult,
+  AiStatus,
+  AiThinkingPhase,
+  AiTokenUsage,
+  AiUsagePhase,
+  ClientAgentBootstrap,
+  ClientAgentToolResult,
+  PlannerAgentDto,
+  AiSourceType,
+} from './aiTypes';
 
-export type AiSourceType = 'rote' | 'article';
-
-export interface AiStatus {
-  enabled: boolean;
-  vectorEnabled: boolean;
-  publicExploreVectorEnabled: boolean;
-  eligible: boolean;
-  available: boolean;
-  memoryAvailable?: boolean;
-  chatAvailable?: boolean;
-  chatAllowed?: boolean;
-  chatProviderId?: string;
-  chatModel?: string;
-  chatMode?: 'disabled' | 'site' | 'local';
-  memoryStats?: {
-    roteCount: number;
-    indexedRoteCount: number;
-  };
-}
-
-export interface AiSemanticResult {
-  id?: string;
-  ownerId?: string;
-  sourceType: AiSourceType;
-  sourceId: string;
-  chunkIndex?: number;
-  text?: string;
-  preview?: string;
-  similarity: number;
-  metadata: {
-    title?: string;
-    tags?: string[];
-    state?: string;
-    archived?: boolean;
-    createdAt?: string;
-    updatedAt?: string;
-    [key: string]: unknown;
-  };
-}
-
-export type AiToolCallingProbeResult = {
-  supported: boolean;
-  message: string;
-  toolName?: string;
-  arguments?: Record<string, unknown>;
-  rawContent?: string;
-  error?: string;
-};
-
-export interface AiProviderTestResult {
-  success: boolean;
-  eligible?: boolean;
-  chatAvailable?: boolean;
-  vectorAvailable?: boolean;
-  model?: string;
-  latencyMs?: number;
-  sample?: string;
-  usage?: AiTokenUsage;
-  toolCalling?: AiToolCallingProbeResult;
-}
-
-export type AiProviderTestProgressStep = 'site' | 'personal_remote' | 'local_chat' | 'tool_calling';
-
-export type AiProviderTestProgressHandler = (step: AiProviderTestProgressStep) => void;
-
-export type LifecycleScope = 'active' | 'archived' | 'all' | 'unspecified';
-export type TaskStatusScope = 'open' | 'closed' | 'all' | 'unspecified';
-
-export interface NormalizedTimeRange {
-  from: string;
-  to: string;
-  label: string;
-}
-
-export interface RetrievalScope {
-  ownerId: string;
-  query: string;
-  tags: string[];
-  excludeTags: string[];
-  semanticScope: string[];
-  sourceTypes: AiSourceType[];
-  timeRange: NormalizedTimeRange | null;
-  lifecycleScope: LifecycleScope;
-  taskStatusScope: TaskStatusScope;
-  limit: number;
-  cursor: string | null;
-  excludeIds: string[];
-}
-
-export interface RetrievalSnippet {
-  id: string;
-  sourceType: AiSourceType;
-  sourceId: string;
-  title?: string;
-  tags?: string[];
-  createdAt?: string;
-  similarity: number;
-  text: string;
-}
-
-export interface RetrievalToolResult {
-  canonicalizedArgs: RetrievalScope;
-  resultCount: number;
-  topSnippets: RetrievalSnippet[];
-  cursor: string | null;
-  warnings: string[];
-}
-
-export interface PlannerDebugTrace {
-  toolCalls: Array<{ step: number; name: string; args: unknown }>;
-  canonicalizedArgs: RetrievalScope[];
-  warnings: string[];
-  probeCounts: number[];
-  finishReason?: string;
-  fallbackReason?: string;
-  providerError?: string;
-  toolError?: string;
-}
-
-export interface PlannerAgentDto {
-  originalMessage: string;
-  retrievalNeeded: boolean;
-  scope: RetrievalScope | null;
-  toolResult: RetrievalToolResult | null;
-  clarification: { question: string; reason?: string } | null;
-  debugTrace: PlannerDebugTrace;
-}
-
-export interface AiClarification {
-  question: string;
-  pendingPlan?: PlannerAgentDto | null;
-}
-
-export type AiAgentPhase =
-  | 'understanding'
-  | 'planning'
-  | 'tool_calling'
-  | 'retrieving'
-  | 'reading'
-  | 'answering';
-
-export type AiAgentToolProgressStatus =
-  | 'determining_scope'
-  | 'retrieving_sources'
-  | 'reading_source'
-  | 'finding_related'
-  | 'loading_tags';
-
-export type AiUsagePhase = 'planning' | 'tool_decision' | 'answer';
-export type AiThinkingPhase =
-  | 'route_decision'
-  | 'evidence_decision'
-  | 'retrieval_planning'
-  | 'answer';
-
-export type AiTokenUsage = {
-  prompt_tokens: number;
-  completion_tokens: number;
-  total_tokens: number;
-};
-
-export interface AiAgentClientState {
-  conversationId?: string;
-  previousPlan?: PlannerAgentDto | null;
-  seenSourceIds?: string[];
-  selectedContext?: {
-    currentRoteId?: string;
-    currentArticleId?: string;
-    selectedSourceIds?: string[];
-    selectedTags?: string[];
-  } | null;
-  clientContext?: AiClientRequestContext | null;
-  stateVersion?: number;
-}
-
-export type AiChatStreamHandlers = {
-  onRunStarted?: (runId: string) => void;
-  onProgress?: (phase: AiAgentPhase) => void;
-  onHeartbeat?: (heartbeat: { phase: AiAgentPhase; seq: number; timestamp: string }) => void;
-  onToolStarted?: (toolName: string, args?: unknown) => void;
-  onToolProgress?: (toolName: string, status: AiAgentToolProgressStatus) => void;
-  onToolFinished?: (toolName: string, summary?: string) => void;
-  onPlan?: (plan: PlannerAgentDto) => void;
-  onClarification?: (clarification: AiClarification) => void;
-  onSources?: (sources: AiSemanticResult[]) => void;
-  onThinking?: (phase: AiThinkingPhase, text: string) => void;
-  onDelta?: (text: string) => void;
-  onStatePatch?: (state: Partial<AiAgentClientState>) => void;
-  onUsage?: (usage: AiTokenUsage, phase: AiUsagePhase) => void;
-  onDone?: () => void;
-  onError?: (message: string) => void;
-};
-
-export type ClientAgentBootstrap = {
-  systemPrompt: string;
-  finalAnswerInstruction: string;
-  tools: Array<{
-    type: 'function';
-    function: {
-      name: string;
-      description: string;
-      parameters: Record<string, unknown>;
-    };
-  }>;
-  policy: {
-    maxIterations: number;
-    maxToolCalls: number;
-    maxSources: number;
-  };
-};
-
-export type ClientAgentToolResult = {
-  observations: string[];
-  displaySummary?: unknown;
-  modelContent: string;
-  sources: AiSemanticResult[];
-  plan?: PlannerAgentDto;
-  statePatch?: Partial<AiAgentClientState>;
-  state: AiAgentClientState;
-  sourceKeys: string[];
-  clarification?: AiClarification;
-};
+export type * from './aiTypes';
 
 export const getAiStatus = () => get('/ai/status').then((res) => res.data as AiStatus);
 
@@ -233,31 +32,6 @@ export const testSiteAiProvider = (onProgress?: AiProviderTestProgressHandler) =
     message: res.message,
   }));
 };
-
-export type AiChatPayload = {
-  message: string;
-  mode?: 'chat' | 'review' | 'organize';
-  limit?: number;
-  pendingPlan?: PlannerAgentDto | null;
-  clarificationAnswer?: string;
-  previousPlan?: PlannerAgentDto | null;
-  excludeIds?: string[];
-  history?: { role: 'user' | 'assistant'; content: string }[];
-  state?: AiAgentClientState | null;
-  clientContext?: AiClientRequestContext | null;
-  debug?: boolean;
-  enableThinking?: boolean;
-};
-
-export interface AiClientRequestContext {
-  nowIso: string;
-  localDate: string;
-  localDateTime: string;
-  timeZone?: string;
-  utcOffsetMinutes: number;
-  locale?: string;
-  calendar?: string;
-}
 
 function padDatePart(value: number): string {
   return String(value).padStart(2, '0');
@@ -306,6 +80,7 @@ export function buildAiClientTimeContextMessage(context: AiClientRequestContext)
     context.calendar ? `Client calendar: ${context.calendar}` : null,
     'Resolve relative date phrases such as today, yesterday, this month, last month, 最近, 本月, and 上月 using this context.',
     'For Rote search tools, prefer structured timeRange preset/rolling/relative_between or pass the original phrase as timeExpression. Use from/to only for explicit absolute dates.',
+    'For broad recent/latest record reviews or recurring-theme analysis, use selection recent with a default limit of 30 and dateField createdAt. Use updatedAt for modification/activity wording. For focused topics, use selection relevance with an explicit time range.',
   ]
     .filter((line): line is string => Boolean(line))
     .join('\n');
@@ -397,17 +172,11 @@ async function readResponseError(response: Response): Promise<string> {
 function parseSseEvent(block: string): { event: string; data: unknown } | null {
   let event = 'message';
   const dataLines: string[] = [];
-
   block.split('\n').forEach((line) => {
-    if (line.startsWith('event:')) {
-      event = line.slice(6).trim();
-    } else if (line.startsWith('data:')) {
-      dataLines.push(line.slice(5).trimStart());
-    }
+    if (line.startsWith('event:')) event = line.slice(6).trim();
+    else if (line.startsWith('data:')) dataLines.push(line.slice(5).trimStart());
   });
-
   if (dataLines.length === 0) return null;
-
   const dataText = dataLines.join('\n');
   try {
     return { event, data: JSON.parse(dataText) };
@@ -438,12 +207,8 @@ async function readAiStreamResponse(
   response: Response,
   handlers: AiChatStreamHandlers
 ): Promise<void> {
-  if (!response.ok) {
-    throw new Error(await readResponseError(response));
-  }
-  if (!response.body) {
-    throw new Error('Streaming response is not supported in this browser');
-  }
+  if (!response.ok) throw new Error(await readResponseError(response));
+  if (!response.body) throw new Error('Streaming response is not supported in this browser');
 
   const reader = response.body.getReader();
   const decoder = new TextDecoder();
@@ -458,7 +223,7 @@ async function readAiStreamResponse(
       const runId = (parsed.data as { runId?: string })?.runId;
       if (runId) handlers.onRunStarted?.(runId);
     } else if (parsed.event === 'progress') {
-      const data = parsed.data as { phase?: AiAgentPhase; message?: string };
+      const data = parsed.data as { phase?: AiAgentPhase };
       if (data.phase) handlers.onProgress?.(data.phase);
     } else if (parsed.event === 'heartbeat') {
       const data = parsed.data as { phase?: AiAgentPhase; seq?: number; timestamp?: string };
@@ -469,14 +234,8 @@ async function readAiStreamResponse(
       const data = parsed.data as { toolName?: string; args?: unknown };
       if (data.toolName) handlers.onToolStarted?.(data.toolName, data.args);
     } else if (parsed.event === 'tool_progress') {
-      const data = parsed.data as {
-        toolName?: string;
-        status?: AiAgentToolProgressStatus;
-        message?: string;
-      };
-      if (data.toolName && data.status) {
-        handlers.onToolProgress?.(data.toolName, data.status);
-      }
+      const data = parsed.data as { toolName?: string; status?: AiAgentToolProgressStatus };
+      if (data.toolName && data.status) handlers.onToolProgress?.(data.toolName, data.status);
     } else if (parsed.event === 'tool_finished') {
       const data = parsed.data as { toolName?: string; summary?: string };
       if (data.toolName) handlers.onToolFinished?.(data.toolName, data.summary);
@@ -485,9 +244,7 @@ async function readAiStreamResponse(
       if (plan) handlers.onPlan?.(plan);
     } else if (parsed.event === 'clarification') {
       const clarification = parsed.data as AiClarification;
-      if (clarification?.question) {
-        handlers.onClarification?.(clarification);
-      }
+      if (clarification?.question) handlers.onClarification?.(clarification);
     } else if (parsed.event === 'sources') {
       const sources = (parsed.data as { sources?: AiSemanticResult[] })?.sources;
       handlers.onSources?.(Array.isArray(sources) ? sources : []);
@@ -509,14 +266,10 @@ async function readAiStreamResponse(
       const state = (parsed.data as { state?: Partial<AiAgentClientState> })?.state;
       if (state) handlers.onStatePatch?.(state);
     } else if (parsed.event === 'usage') {
-      const data = parsed.data as {
-        phase?: AiUsagePhase;
-        usage?: Partial<AiTokenUsage>;
-      };
+      const data = parsed.data as { phase?: AiUsagePhase; usage?: Partial<AiTokenUsage> };
       const rawUsage = data.usage;
-      const phase = data.phase;
       if (
-        phase &&
+        data.phase &&
         rawUsage &&
         typeof rawUsage.prompt_tokens === 'number' &&
         typeof rawUsage.completion_tokens === 'number' &&
@@ -528,7 +281,7 @@ async function readAiStreamResponse(
             completion_tokens: rawUsage.completion_tokens,
             total_tokens: rawUsage.total_tokens,
           },
-          phase
+          data.phase
         );
       }
     } else if (parsed.event === 'done') {
@@ -545,9 +298,7 @@ async function readAiStreamResponse(
     while (true) {
       const { done, value } = await reader.read();
       if (done) break;
-
       buffer += decoder.decode(value, { stream: true }).replace(/\r\n/g, '\n');
-
       let boundary = buffer.indexOf('\n\n');
       while (boundary !== -1) {
         const block = buffer.slice(0, boundary);
@@ -556,7 +307,6 @@ async function readAiStreamResponse(
         boundary = buffer.indexOf('\n\n');
       }
     }
-
     const tail = buffer.trim();
     if (tail) dispatch(tail);
     if (!doneReceived) handlers.onDone?.();
